+++ lib/ash_age.ex
@@ -11,7 +11,7 @@ defmodule AshAge do
       defmodule MyApp.PostgrexTypes do
         Postgrex.Types.define(
           MyApp.PostgrexTypes,
-          [AshAge.Type.Agtype.Extension] ++ Ecto.Adapters.Postgres.extensions(),
+          [AshAge.Postgrex.AgtypeExtension] ++ Ecto.Adapters.Postgres.extensions(),
           []
         )
       end